finished db configuration

+++ net/App.Data/Entities/Movie.cs
@@ -7,5 +7,5 @@ public class Movie : IBaseEntity<string>
     public string Id { get; set; }
     public string Name { get; set; }
     public DateTime ProductionYear { get; set; }
-    public decimal BoxOffice { get; set; }
+    public int BoxOffice { get; set; }
 }
